+++ packages/circuit-ui/components/Button/Button.tsx
@@ -35,7 +35,7 @@ import {
 } from '../../styles/style-mixins';
 import { ReturnType } from '../../types/return-type';
 import { useComponents } from '../ComponentsContext';
-import useClickHandler from '../../hooks/use-click-handler';
+import { useClickHandler } from '../../hooks/useClickHandler';
 
 export interface BaseProps {
   'children': ReactNode;
