+++ packages/circuit-ui/components/RadioButton/RadioButton.tsx
@@ -24,7 +24,7 @@ import {
   focusOutline,
 } from '../../styles/style-mixins';
 import { uniqueId } from '../../util/id';
-import useClickHandler from '../../hooks/use-click-handler';
+import { useClickHandler } from '../../hooks/useClickHandler';
 
 export interface RadioButtonProps extends HTMLProps<HTMLInputElement> {
   /**
