+++ packages/circuit-ui/components/Step/Step.spec.js
@@ -14,9 +14,9 @@
  */
 
 import Step from './Step';
-import useStep from './hooks/use-step';
+import { useStep } from './hooks/useStep';
 
-jest.mock('./hooks/use-step', () => jest.fn(() => ({})));
+jest.mock('./hooks/useStep', () => ({ useStep: jest.fn(() => ({})) }));
 
 describe('Step', () => {
   afterAll(() => {
