+++ packages/circuit-ui/components/Tag/Tag.tsx
@@ -20,7 +20,7 @@ import { Theme } from '@sumup/design-tokens';
 
 import styled, { StyleProps } from '../../styles/styled';
 import { typography, focusOutline } from '../../styles/style-mixins';
-import useClickHandler from '../../hooks/use-click-handler';
+import { useClickHandler } from '../../hooks/useClickHandler';
 import { CloseButton, CloseButtonProps } from '../CloseButton/CloseButton';
 
 type BaseProps = {
